backend/requirements.txt
# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Web framework
fastapi==0.104.1
mangum==0.17.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image processing
Pillow==10.1.0
exifread==3.0.0

# AI/ML - Google Generative AI
google-generativeai==0.3.2

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 